pyproject.toml
[project]
name = "krm3"
version = "0.19.0"
description = "KRM3"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "django<5.2",
    "django-admin-rangefilter>=0.13.2",
    "django-adminfilters>=2.6",
    "django-click>=2.4.1",
    "django-constance>=4.3.2",
    "django-cors-headers>=4.7.0",
    "django-crispy-forms>=2.3",
    "django-debug-toolbar>=5.1.0",
    "django-filter>=25.1",
    "django-mptt-admin>=2.7.2",
    "django-picklefield>=3.3",
    "django-qr-code>=4.1.0",
    "django-regex>=0.5.0",
    "django-tables2>=2.7.5",
    "djangorestframework-camel-case>=1.4.2",
    "djoser>=2.3.1",
    "docutils>=0.21.2",
    "drf-spectacular>=0.28.0",
    "markdown>=3.7",
    "natural-keys>=2.1.1",
    "parametrize>=0.1.1",
    "psycopg[binary,pool]>=3.2.6",
    "pyoxr>=1.6",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sentry-sdk>=2.25.1",
    "tabulate>=0.9.0",
    "django-smart-admin[full]>=2.6.0",
    "tablib[xlsx]>=3.8.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
]

[dependency-groups]
dev = [
    "bumpversion>=0.6.0",
    "detect-secrets>=1.5.0",
    "django-extensions>=3.2.3",
    "django-factory-boy>=1.0.0",
    "django-stubs>=5.1.3",
    "django-webtest>=1.9.13",
    "djangorestframework-stubs>=3.15.3",
    "opencv-python>=4.11.0.86",
    "pdbpp>=0.10.3",
    "pre-commit>=4.2.0",
    "pydocstyle>=6.3.0",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-dotenv>=0.5.2",
    "pytest-env>=1.1.5",
    "pytest-param>=0.1.1",
    "pytest-race>=0.2.0",
    "responses>=0.25.7",
]
image = [
    "circus>=0.19.0",
    "opencv-python-headless>=4.11.0.86",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
